MongoDb and Python
These files are the result of following a MongoDb and Python tutorial.  

PyMongoTest.py is testing a connection to MongoDb.
createsamples.py creates a business database and a reviews collection and randomly generates content for that collection.
exploreMongoDb.py explores basic PyMongo functionality such as searching, altering, and deleting entries.  (or Find, Update, and Delete).